codebase1/FaceRecognition/requirements.txt
numpy
opencv-python
opencv-contrib-python
mysql-connector-python
pyttsx3
Pillow
pysimplegui